Clone priv rep

+++ management_server/src/main/java/com/IoTSim/management_server/context/user/service/UserService.java
@@ -1,24 +1,23 @@
 package com.IoTSim.management_server.context.user.service;
 
 import com.IoTSim.management_server.api.exceptions.UserNotFoundException;
+
+import com.IoTSim.management_server.context.user.api.UserInfoResponse;
 import com.IoTSim.management_server.context.user.dto.UserDto;
 import com.IoTSim.management_server.context.user.mapper.UserMapper;
-import com.IoTSim.management_server.context.user.model.Role;
-import com.IoTSim.management_server.models.ConfirmationToken;
 import com.IoTSim.management_server.context.user.model.User;
-import com.IoTSim.management_server.repository.ConfirmationTokenRepository;
 import com.IoTSim.management_server.context.user.repository.UserRepository;
-import com.IoTSim.management_server.service.EmailService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
-import org.springframework.http.ResponseEntity;
+import com.IoTSim.management_server.context.user.model.Role;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import org.springframework.util.ObjectUtils;
 
 import java.util.*;
@@ -28,35 +27,29 @@ import java.util.*;
 public class UserService implements UserDetailsService {
 
     private final UserRepository userRepository;
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-    private final EmailService emailService;
-    private final ConfirmationTokenRepository confirmationTokenRepository;
     private final UserMapper mapper;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<User> user = userRepository.findByUsername(username);
 
-        if (user.isEmpty())
-            throw new UsernameNotFoundException("User not found");
-
-        return user.get();
+        return userRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
     }
 
-    public List<UserDto> getAllUsers(){
+    @Transactional
+    public List<UserInfoResponse> getAllUsers(){
         List<User> userList = userRepository.findAll();
-        return mapper.UserListToUserDtoList(userList);
+        return mapper.UserListToUserInfoResponseList(userList);
     }
 
+    @Transactional
     public void deleteUserById(Long id){
         User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if (ObjectUtils.isEmpty(user))
-            throw new UserNotFoundException("Current user not found");
-        if (userRepository.existById(id))
-            throw new UserNotFoundException("User not found");
+
+        if (ObjectUtils.isEmpty(user) || !userRepository.existById(id))
+            throw new UserNotFoundException();
         if (
-                user.getId() == id ||
-                        user.getRole().equals(Role.ADMIN)
+                Objects.equals(user.getId(), id)
         ) {
             userRepository.deleteById(id);
         } else {
@@ -78,15 +71,13 @@ public class UserService implements UserDetailsService {
         return ResponseEntity.badRequest().body("Error: Couldn't verify email");
     }
 */
+    @Transactional
     public void updateUser(UserDto userDto){
         User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if (ObjectUtils.isEmpty(user))
-            throw new UserNotFoundException("Current user not found");
-        if (userRepository.existById(userDto.getId()))
-            throw new UserNotFoundException("User not found");
+        if (ObjectUtils.isEmpty(user) || !userRepository.existById(userDto.getId()))
+            throw new UserNotFoundException();
         if (
-                user.getId() == userDto.getId() ||
-                        user.getRole().equals(Role.ADMIN)
+                Objects.equals(user.getId(), userDto.getId())
         ) {
             User savingUser = mapper.UserDtoToUser(userDto);
             userRepository.save(savingUser);
@@ -105,17 +96,18 @@ public class UserService implements UserDetailsService {
                 .orElseThrow(() -> new UsernameNotFoundException("User not found"));
     }
 
-    public UserDto getUserById(Long id){
+    @Transactional
+    public UserInfoResponse getUserById(Long id){
         User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         if (ObjectUtils.isEmpty(user))
-            throw new UserNotFoundException("Current user not found");
-        if (userRepository.existById(id))
-            throw new UserNotFoundException("User not found");
+            throw new UserNotFoundException();
         if (
-                user.getId() == id ||
-                        user.getRole().equals(Role.ADMIN)
+                Objects.equals(user.getId(), id)
         ) {
-            return mapper.UserToUserDto(userRepository.findById(id).get());
+            return mapper.UserToUserInfoResponse(
+                    userRepository.findById(id)
+                            .orElseThrow(UserNotFoundException::new)
+            );
         } else {
             throw new AccessDeniedException("Access Denied");
         }
